refactor: move PercentileLatencyComponent to new directory
moves PercentileLatencyComponent to the new PercentileLatency directory and updates all associated imports
