+++ client/src/containers/CloudWatchMetrics/Throttle/ThrottleComponent.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import ThrottleComponent from "../../../components/ThrottleComponent";
+import ThrottleComponent from "./ThrottleComponent";
 import { Line } from "react-chartjs-2";
 
 // Mock the Line component to avoid rendering an actual chart during tests.
