+++ client/src/containers/CloudWatchMetrics/Throttle/ThrottleComponent.tsx
