+++ client/src/containers/CloudWatchMetrics/TotalDuration/TotalDurationComponent.tsx
